refactor: Remove content editor and move functions
