--- src/components/Admin/ContentEditor.js
@@ -1,41 +1,0 @@
-import React, { Component } from "react"
-import CKEditor from "@ckeditor/ckeditor5-react"
-import ClassicEditor from "@ckeditor/ckeditor5-build-classic"
-import UploadAdapter from "./UploadAdapter"
-
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <h2>Using CKEditor 5 build in React</h2>
-        <CKEditor
-          editor={ClassicEditor}
-          data="<p>Hello from CKEditor 5!</p>"
-          onInit={editor => {
-            editor.plugins.get(
-              "FileRepository"
-            ).createUploadAdapter = loader => {
-              return new UploadAdapter(loader)
-            }
-
-            editor.plugins.get("ImageResize")
-
-            console.log("Editor is ready to use!", editor)
-          }}
-          onChange={(event, editor) => {
-            const data = editor.getData()
-            console.log({ event, editor, data })
-          }}
-          onBlur={(event, editor) => {
-            console.log("Blur.", editor)
-          }}
-          onFocus={(event, editor) => {
-            console.log("Focus.", editor)
-          }}
-        />
-      </div>
-    )
-  }
-}
-
-export default App
